[performance] Cache geometry data from DB to prevent query-in-a-loop madness.
Before this, geometry data (of which there isn't much), is asked for in
a loop for each incident (of which there can be many).

This patch grabs all of the geometry data the first time any of it is
asked for, and caches in a lookup table within the app, saving thousands
of pointless queries.