Closes #2190
Depending on if you request all the categories or if you request a category by ID the json returns different keys to index the title, color, and description fields. This just confuses things and is unnecessary. I'm going to set the indexes to match the names used in the database:
category_title
category_description
category_color