Rewrite of mapping code - ushahidi.js
* NOTE: This is a major change so hold off updating a live deployment
from this branch for the next couple of hours as the changes get tested

* timeline.js (formerly used to rendering the timeline on the landing
page) has been replaced with ushahidi.js. The latter handles all map
rendering functions and also updates to the overlays when the zoom
level changes, report filters (date, category, media type) etc are
applied. The new script exposes an API for registering callback
functions (JavaScript functions) to be called when report filters for
the map are upped. This change effectively decouples the map from the
timeline. Finally.

* Added jqplot libraries for rendering the timeline but a couple of
modifications are still required so that the libraries are loaded only
when a 'show timeline' setting is enabled.

* The view files have been grouped into directories to reflect the
controllers associated with them. Therefore all the view files for the
alerts controller are in "views/alerts", the ones for reports are in
"views/reports/". The same scheme has been used in the view files for
the admin section i.e. "views/admin/manage/layers" for the views used
for managing layers(overlays), "views/admin/reports" for the admin
reports controller etc