Overlay KMLs on main map
* The KML listed on the sidebar ("other layers") can now be overlaid on
the map. This feature has always been there though previously, the name
of the layer that was displayed on the layer switcher control was not
the same as the name of the selected layer. This has now been rectified.

* Added an extra event 'delete layer' in ushahidi.js which triggers
removal of a layer from the map. This event is registered as soon as
the map is initialized so that when invoked, the 'trigger' method does
not exit prematurely - what happens when an event is not found in the
callbacks registry. The name of the layer to be deleted is the second
parameter when calling the 'trigger' method

* Coding style updates for the layer method in the json controller