Apply date and media filters to the timeline
* The timeline now responds to the date filters (start and end dates)
and media filters. Also, the minimum no. of data points required for
the timeline has been set to two - for some reason, jqplot hangs when
there is only one data point

* I have also removed the "play" button (previously used to visualize
the influx of reports into a deployment) primarily because we are no
longer using timeline.js. This feature may however be re-added later on
as a plugin.