Add radius layer functionality to ushahidi.js & cleanup settings JS
* The JS for the alerts page now sits on top of ushahidi.js. There is
now an updateAlertCoordinates function that is subscribed to the
'markerpositionchanged' event so that the alert coordinates (lat/lon)
can be updated on the alerts subscription form each time alert center
shifts

* The background of the popups is now preloaded when the map is
initialized as opposed to when the feature is selected. This ensures
that it is cached in advance; before adding any layers on the map

* Renamed the "REPORTS" layer type to GEOJSON so that it reflects the
type of data associated with the layer; mnemonics. In line with this
norm, the "defaultMarkerStyle" property has also been renamed to
"GeoJSONStyle"

* Cleaned up the settings_js file and delegated the map functionality
to ushahidi.js. This required adding extra API methods to handle
adjustment of the zoom level via the slider, change of default country
and base layer