Accommodate rendering of checkins in the new mapping script
* The main_js file is now much cleaner - devoid of OpenLayers code; all
this is in ushahidi.js. The data from the checkins API call is now
passed onto a callback function in main_js that formats it into a JSON
object which is then parsed into an array of OpenLayer.Feature.Vector
objects by the addLayer function ins ushahidi.js

* The checkins API - for fetching checkins - has been refactored so
that the each checkin entry includes a user object that contains all
the user's data (name, color etc). It doesn't make much sense to split
the payload between checkins and users

* I've removed the checkins listing from the sidebar on the main page.
The plan is to implement that listing (among others) using Backbone JS