Use key-value storage for the settings table
* The settings table is now using key-value system to store data. The
key MUST be unique so repetitions are not allowed - a UNIQUE index on
this column ensures this.

* The upgrade script (upgrade90-91.sql)  to effect the change is also
included in this commit. The script first creates the "new_settings"
table and creates entries by selecting individual columns from the old
settings table. Thereafter, the old settings table is deleted and the
new one renamed to "settings"

* The settings model class now has convenience methods - mostly static
ones - that provide a simple interface for interacting with the
restructured settings table