Use new settings table in the installer. Closes #592
* I've whittled down the installer  to 3 main scripts - the bootstrap
(index.php), wizard (wizard.php) and the utilities (utils.php). The
views for each of the install stages are in "installer/pages/". The
entire installation process is managed by "wizard.php" - keeps track of
the current stage in the installation process thereby allowing a user
to resume their session in case they accidentally close a tab.

* The "Base Path" field has been removed from the database info screen.
We already auto-detect this so there's no need to ask the user to
specify it.

* When using "Advanced Mode", the installer no-longer prompts for an
API key when OSM is selected as the default map provider. Google and
Bing however require a map API key to be provided.

* The salt pattern is now set by the installer. Turns out that the
numbers in the pattern have to be sorted in ascending order. Closes #494