Fix retrieve cities problem. Closes #604
* The settings controller was using an old Geonames API endpoint and
therefore the city table was not being populated with the name of the
cities. Geonames changed the web services endpoint to
http://api.geonames.org. The new endpoint requires a valid Geonames
username which must be "enabled for the free web services". The maximum
no. of cities to be retrievable has been restricted to 1000 though we
ought to make this item a configurable parameter - shelved for a later
release.

Reported by: @levental