Move installation check to index.php
* It makes more sense to perform the installation check in index.php;
before the Kohana framework is bootstrapped. In light of this change,
config.php has to be created from the provided template and has also
been added as entry in .gitignore. The same goes for encryption.php and
auth.php - these files are now automatically created by the installer.
Consequently, the "requirements" view in the installer has also been
updated to reflect this change

* The v.93-94 DB upgrade seems to have left out the changes from to the
main setup script (ushahidi.sql)