added Web.config file to allow one to host their uhsahidi installationd on Windows servers running under IIS
This change will allow you to host your ushahidi  installation on
Windows Azure. This was not possible before and you could not host your
ushahidi installation on a Windows web server running IIS as the web
server because the .htaccess file only works on Apache and will not work
under IIS.
IIS howver will only parse the rewrite rules in a Web.config file.
Here I have put all the rewrite rules in the web.config file so that I
can be able to host my Ushahidi installation on Windows Azure.