Updating upgrade SQL to delete viddler
New instalations won't get viddler on the add-on screen since the plugin
directory is gone. However, existing instalations need to remove it from
the table to avoid a nasty empty line on Add-on screen.

Not sure if we should be checking if the plugin is enabled before doing
this. But then, if the plugin is enabled, the whole thing will break
miserably anyway because its code is gone. Thought I should mention this
as I'm not sure how keen we are on keeping retro-compatibility. :)