Fixing timeline js timing out (#483)
Issue if being cause by selectToUISlider() plugin. It calls a
jQuery::width() on each of the elements it creates for the month names (it
can go up to quite a few on large time spans)

Updated function to call jQuery::width() only on the first element of the
set, and letting jQuery handle the css() call for the whole collection. I
suspect it will use each() internally, but I also suspect it will have
smarts to do it more efficiently.